Render login prompt rather than auth redirect
Due to `CVE-2015-9284` OmniAuth 2 does not allow `GET` requests as part
of the `Strategy#request_phase`. The recommendation is to instead change
calls to action to "log in" to [`POST` a form with an authenticity token
from
rack-protection](https://github.com/omniauth/omniauth/wiki/Upgrading-to-2.0#racksinatra).

This all makes sense from the perspective of `OmniAuth` and for closing
the XSRF hole. But it doesn't fit with how `heroku-bouncer` works today
wherein a request will be silently redirected into the OAuth flow.

This commit changes that behavior such that we redirect/render an
extremely basic login form. We've added a new `:login_path` option
allowing users to specify a path to redirect to. From there they need to
render their own `<form>` with the correct inputs, `action`, etc…, as
explained in the README.